min cluster must be int

+++ Pylette/src/extractors/hdbscan.py
@@ -25,7 +25,7 @@ class HDBSCANExtractor(ColorExtractorBase):
         from sklearn.cluster import HDBSCAN
 
         arr = np.squeeze(arr)
-        clusterer = HDBSCAN(min_cluster_size=len(arr) / 10, allow_single_cluster=True) # no cluster with less than 10% of points
+        clusterer = HDBSCAN(min_cluster_size=int(len(arr) / 10), allow_single_cluster=True) # no cluster with less than 10% of points
         labels = clusterer.fit_predict(arr)
         unique_labels, counts = np.unique(labels, return_counts=True)
         label_freq = counts / counts.sum()
